fix: actually write changes on "superset import-datasources"
When running the following command:

    superset import-datasources --path=datasource.json

the data stored in "datasources.json" may either be dict or a list. In the case
of a dataset exported from a live superset instance, it is a list. When it is
re-imported, the `SqlaTable.import_from_dict` utility method is used. This
method does not commit changes to the database, so the imported datasets are
never saved to the database, causing frustration and sadness.